Converted all benchmarks from double to single precision floats.

+++ src/Benchmarks/SharpNeat.Benchmarks/Neat/DistanceMetrics/DistanceMetricUtilsBenchmarks.cs
@@ -7,19 +7,19 @@ namespace SharpNeat.Neat.DistanceMetrics;
 
 public class DistanceMetricUtilsBenchmarks
 {
-    readonly ConnectionGenes<double>[] _genomeArr;
+    readonly ConnectionGenes<float>[] _genomeArr;
 
     public DistanceMetricUtilsBenchmarks()
     {
-        var metaNeatGenome = MetaNeatGenome<double>.CreateAcyclic(12, 1, new LeakyReLU<double>());
-        var popLoader = new NeatPopulationLoader<double>(metaNeatGenome);
-        List<NeatGenome<double>> genomeList = popLoader.LoadFromZipArchive("data/binary11.pop");
+        var metaNeatGenome = MetaNeatGenome<float>.CreateAcyclic(12, 1, new LeakyReLU<float>());
+        var popLoader = new NeatPopulationLoader<float>(metaNeatGenome);
+        List<NeatGenome<float>> genomeList = popLoader.LoadFromZipArchive("data/binary11.pop");
         _genomeArr = genomeList.Select(x => x.ConnectionGenes).ToArray();
     }
 
     [Benchmark]
     public void CalculateEuclideanCentroid()
     {
-        _ = DistanceMetricUtils.CalculateEuclideanCentroid((ReadOnlySpan<ConnectionGenes<double>>)_genomeArr);
+        _ = DistanceMetricUtils.CalculateEuclideanCentroid((ReadOnlySpan<ConnectionGenes<float>>)_genomeArr);
     }
 }
