+++ src/Benchmarks/SharpNeat.Benchmarks/NeuralNets/ActivationFunctions/Vectorized/ActivationFunctionsBenchmarks.cs
@@ -1,6 +1,6 @@
 ﻿using System.Runtime.InteropServices;
 using BenchmarkDotNet.Attributes;
-using Redzen.Numerics.Distributions.Double;
+using Redzen.Numerics.Distributions.Float;
 
 namespace SharpNeat.NeuralNets.ActivationFunctions.Vectorized;
 
@@ -9,24 +9,24 @@ namespace SharpNeat.NeuralNets.ActivationFunctions.Vectorized;
 /// </summary>
 public class ActivationFunctionsBenchmarks
 {
-    static readonly IActivationFunction<double> __LeakyReLU = new LeakyReLU<double>();
-    static readonly IActivationFunction<double> __LeakyReLUShifted = new LeakyReLUShifted<double>();
-    static readonly IActivationFunction<double> __MaxMinusOne = new MaxMinusOne<double>();
-    static readonly IActivationFunction<double> __PolynomialApproximantSteep = new PolynomialApproximantSteep<double>();
-    static readonly IActivationFunction<double> __QuadraticSigmoid = new QuadraticSigmoid<double>();
-    static readonly IActivationFunction<double> __ReLU = new ReLU<double>();
-    static readonly IActivationFunction<double> __SoftSignSteep = new SoftSignSteep<double>();
-    static readonly IActivationFunction<double> __SReLU = new SReLU<double>();
-    static readonly IActivationFunction<double> __SReLUShifted = new SReLUShifted<double>();
+    static readonly IActivationFunction<float> __LeakyReLU = new LeakyReLU<float>();
+    static readonly IActivationFunction<float> __LeakyReLUShifted = new LeakyReLUShifted<float>();
+    static readonly IActivationFunction<float> __MaxMinusOne = new MaxMinusOne<float>();
+    static readonly IActivationFunction<float> __PolynomialApproximantSteep = new PolynomialApproximantSteep<float>();
+    static readonly IActivationFunction<float> __QuadraticSigmoid = new QuadraticSigmoid<float>();
+    static readonly IActivationFunction<float> __ReLU = new ReLU<float>();
+    static readonly IActivationFunction<float> __SoftSignSteep = new SoftSignSteep<float>();
+    static readonly IActivationFunction<float> __SReLU = new SReLU<float>();
+    static readonly IActivationFunction<float> __SReLUShifted = new SReLUShifted<float>();
 
     const int __loops = 1000;
-    readonly double[] _x = new double[1003];
-    readonly double[] _w = new double[1003];
+    readonly float[] _x = new float[1003];
+    readonly float[] _w = new float[1003];
 
     public ActivationFunctionsBenchmarks()
     {
         // Create some random Gaussian values as the inputs to the activation functions.
-        var gaussian = new ZigguratGaussianSampler(0.0, 2.0, 0);
+        var gaussian = new ZigguratGaussianSampler(0f, 2f, 0);
         for(int i=0; i < _x.Length; i++)
             _x[i] = gaussian.Sample();
     }
@@ -85,10 +85,10 @@ public class ActivationFunctionsBenchmarks
         RunBenchmark(__SReLUShifted);
     }
 
-    private void RunBenchmark(IActivationFunction<double> actFn)
+    private void RunBenchmark(IActivationFunction<float> actFn)
     {
-        ref double xref = ref MemoryMarshal.GetReference(_x.AsSpan());
-        ref double wref = ref MemoryMarshal.GetReference(_w.AsSpan());
+        ref float xref = ref MemoryMarshal.GetReference(_x.AsSpan());
+        ref float wref = ref MemoryMarshal.GetReference(_w.AsSpan());
         int len = _x.Length;
 
         for(int i=0; i < __loops; i++)
