+++ src/Benchmarks/SharpNeat.Benchmarks/NeuralNets/NeuralNetAcyclicBenchmarks.cs
@@ -10,20 +10,20 @@ namespace SharpNeat.NeuralNets;
 
 public class NeuralNetAcyclicBenchmarks
 {
-    static readonly NeuralNetAcyclic<double> __nn;
+    static readonly NeuralNetAcyclic<float> __nn;
 
     static NeuralNetAcyclicBenchmarks()
     {
         // Load neural net model from file, and convert into a neural net instance.
         NetFileModel netFileModel = NetFile.Load("data/genomes/binary11.net");
-        __nn = (NeuralNetAcyclic<double>)NeuralNetConverter.ToNeuralNet<double>(netFileModel);
+        __nn = (NeuralNetAcyclic<float>)NeuralNetConverter.ToNeuralNet<float>(netFileModel);
 
         // Set some non-zero random input values.
         var rng = RandomDefaults.CreateRandomSource();
         var inputs = __nn.Inputs.Span;
         for(int i = 0; i < inputs.Length; i++)
         {
-            inputs[i] = rng.NextDouble();
+            inputs[i] = rng.NextFloat();
         }
     }
