+++ src/Benchmarks/SharpNeat.Benchmarks/NeuralNets/Vectorized/NeuralNetAcyclicBenchmarks.cs
@@ -10,13 +10,13 @@ namespace SharpNeat.NeuralNets.Vectorized;
 
 public class NeuralNetAcyclicBenchmarks
 {
-    static readonly NeuralNetAcyclic<double> __nn;
+    static readonly NeuralNetAcyclic<float> __nn;
 
     static NeuralNetAcyclicBenchmarks()
     {
         // Load neural net model from file, and convert into a neural net instance.
         NetFileModel netFileModel = NetFile.Load("data/genomes/binary11.net");
-        __nn = (NeuralNetAcyclic<double>)NeuralNetConverter.ToNeuralNet<double>(
+        __nn = (NeuralNetAcyclic<float>)NeuralNetConverter.ToNeuralNet<float>(
             netFileModel,
             enableHardwareAcceleratedNeuralNets: true,
             enableHardwareAcceleratedActivationFunctions: false);
@@ -26,7 +26,7 @@ public class NeuralNetAcyclicBenchmarks
         var inputs = __nn.Inputs.Span;
         for(int i = 0; i < inputs.Length; i++)
         {
-            inputs[i] = rng.NextDouble();
+            inputs[i] = rng.NextFloat();
         }
     }
